course_api/requirements.txt
asgiref==3.8.1
distlib==0.3.8
# Django==5.1
Django==4.2.3
# djangorestframework==3.15.2
djangorestframework==3.14.0
filelock==3.13.1
htmlmin==0.1.12
numpy==1.26.4
pandas==2.2.1
platformdirs==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.25.1
django-cors-headers==4.4.0
